docs: add swagger docs & test: add unit tests

+++ src/modules/favorites/services/Favorite.service.ts
@@ -1,7 +1,6 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { FavoriteRepository } from '../repositories/Favorite.repository';
-import { ProductService } from 'src/modules/products/services/Product.service';
-import { ServiceWithProduct } from '../types/ServiceWithProduct.type';
+import { ProductService } from '../../products/services/Product.service';
 import { Favorite } from '../entities/Favorite.entity';
 import { CreateFavoriteDto } from '../dtos/CreateFavorite.dto';
 
@@ -9,34 +8,46 @@ import { CreateFavoriteDto } from '../dtos/CreateFavorite.dto';
 export class FavoriteService {
   constructor(
     private readonly repository: FavoriteRepository,
-    private readonly productService: ProductService
+    private readonly productService: ProductService,
   ) {}
 
-  async createFavorite(clientId, data: CreateFavoriteDto){
-    const favoriteExists = await this.repository.findByServiceAndProduct(data.serviceId, data.productId);
-    if(favoriteExists){
-      throw new BadRequestException("Este produto já é seu favorito!")
-    }
-    await this.productService.verifyProductExists(data.serviceId, data.productId)
-    
-    return this.repository.create({
-      clientId,
-      serviceId: data.serviceId,
-      productId: data.productId,
-      createdAt: new Date()
-    })
+  async createFavorite(clientId: string, data: CreateFavoriteDto) {
+    await this.ensureFavoriteDoesNotExist(clientId, data);
+
+    const product = await this.verifyProduct(data);
+
+    const favorite = await this.saveFavorite(clientId, data);
+
+    return this.buildFavoriteResponse(product, favorite.id);
   }
 
-  async deleteFavorite(id){
-    await this.repository.delete(id);
-  }
-
-  async getAllFavoritesByClient(clientId: string){
+  async getAllFavoritesByClient(clientId: string) {
     const favorites = await this.repository.findByClient(clientId);
     const services = this.getLatestFavoritesByService(favorites);
+    const products = await this.fetchProductsForServices(services);
+    const filteredProducts = this.filterProducts(favorites, products);
+    return filteredProducts;
   }
 
-  private getLatestFavoritesByService(favorites: Favorite[]): { url: string; productId: string }[] {
+  async getFavoriteById(id:string){
+    const favorite = await this.repository.find(id);
+    if(!favorite) throw new NotFoundException("Favorite not Found!")
+    return favorite;
+  }
+
+  async deleteFavorite(id, userId) {
+    const favorite = await this.getFavoriteById(id);
+    if(favorite.client.id === userId){
+      await this.repository.delete(id);
+      return;
+    }
+    throw new ForbiddenException("Você não tem permissão para deletar este favorito.")
+  }
+
+  //Funções auxiliares do GET
+  private getLatestFavoritesByService(
+    favorites: Favorite[],
+  ): { url: string; productId: string }[] {
     const grouped = new Map<string, { url: string; productId: string }>();
 
     for (const fav of favorites) {
@@ -52,14 +63,62 @@ export class FavoriteService {
     return Array.from(grouped.values());
   }
 
-  private async fetchProductsForServices(services: { url: string; productId: string }[]) {
-    const promises = services.map(svc =>
-      this.productService.getProducts(svc.url, Number(svc.productId), 1)
+  private async fetchProductsForServices(
+    services: { url: string; productId: string }[],
+  ) {
+    const promises = services.map((svc) =>
+      this.productService.getProductByUrl(svc.url, Number(svc.productId), 1),
     );
 
     const results = await Promise.all(promises);
 
-    return results.flat(); 
+    return results.flat();
   }
 
+  private filterProducts(favorites: Favorite[], products: any[]) {
+    const favoriteProductIds = new Set(favorites.map((fav) => fav.productId));
+
+    return products.filter((product) =>
+      favoriteProductIds.has(String(product.id)),
+    );
+  }
+  
+  private async ensureFavoriteDoesNotExist(clientId: string, data: CreateFavoriteDto) {
+    const exists = await this.repository.findClientFavoriteExists(
+      data.serviceId,
+      data.productId,
+      clientId,
+    );
+
+    if (exists) {
+      throw new BadRequestException('Este produto já é seu favorito!');
+    }
+  }
+
+  //Funções auxiliares do POST
+
+  private async verifyProduct(data: CreateFavoriteDto) {
+    return this.productService.verifyProductExists(
+      data.serviceId,
+      data.productId,
+    );
+  }
+
+  private async saveFavorite(clientId: string, data: CreateFavoriteDto) {
+    return this.repository.create({
+      clientId,
+      serviceId: data.serviceId,
+      productId: data.productId,
+      createdAt: new Date(),
+    });
+  }
+
+  private buildFavoriteResponse(product: any, favoriteId: string) {
+    return {
+      ...product,
+      favoriteId,
+    };
+  }
+
+  
 }
